third_party/blink/web_tests/external/wpt/digital-credentials/mdoc/mixed-requests.https-expected.txt

This is a testharness.js-based test.
[FAIL] All org-iso-mdoc requests invalid: API rejects (no valid request survives).
  assert_unreached: Should have rejected: undefined Reached unreachable code
[FAIL] Failed request does not poison the coordinator state for subsequent valid requests.
  assert_unreached: Should have rejected: undefined Reached unreachable code
[FAIL] Numeric values in deviceRequest/encryptionInfo are coerced to strings and handled gracefully.
  assert_true: Numeric coercion handled gracefully expected true got false
[FAIL] Object/Array values in deviceRequest/encryptionInfo are handled gracefully.
  assert_true: Object coercion handled gracefully expected true got false
Harness: the test ran to completion.
